Avatar rewritten to TS
