--- src/components/Avatar/index.js
@@ -1,1 +1,0 @@
-export { default } from './Avatar'
